+++ servers/src/common/adapters.rs
@@ -767,7 +767,7 @@ where
 			let _ = tx_pool.reconcile_block(b);
 
 			// First "age out" any old txs in the reorg_cache.
-			let cutoff = Utc::now() - Duration::minutes(tx_pool.config.reorg_cache_period);
+			let cutoff = Utc::now() - Duration::minutes(tx_pool.config.reorg_cache_period as i64);
 			tx_pool.truncate_reorg_cache(cutoff);
 		}
